+++ keyboards/ferris/keymaps/hbmorrisonv5/config.h
@@ -16,41 +16,39 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #pragma once
-
 // Set the mouse settings to a comfortable speed/accuracy trade-off,
 // assuming a screen refresh rate of 60 Htz or higher
 // The default is 50. This makes the mouse ~3 times faster and more accurate
 #define MOUSEKEY_INTERVAL 16
-
 // The default is 20. Since we made the mouse about 3 times faster with the previous setting,
 // give it more time to accelerate to max speed to retain precise control over short distances.
 #define MOUSEKEY_TIME_TO_MAX 40
-
 // The default is 300. Let's try and make this as low as possible while keeping the cursor responsive
 #define MOUSEKEY_DELAY 100
-
 // It makes sense to use the same delay for the mouseweel
 #define MOUSEKEY_WHEEL_DELAY 100
-
 // The default is 100
 #define MOUSEKEY_WHEEL_INTERVAL 50
-
 // The default is 40
 #define MOUSEKEY_WHEEL_TIME_TO_MAX 100
 
-// Default tapping term
+// Pick good defaults for enabling homerow modifiers
 #define TAPPING_TERM 190
-#define PERMISSIVE_HOLD
 #define TAPPING_TERM_PER_KEY
-
 // Tapping term for the paste key
 #define TAPPING_TERM_PASTE 160
+// Tapping term for the leader key
+#define TAPPING_TERM_LEADER 320
 
-// Tapping term for the homerow mod keys
-#define TAPPING_TERM_MODS 1200
+// Permissive hold does not work well with leader key
+// #define PERMISSIVE_HOLD
 
 // Enable key repeat - disabled for now
 // #define QUICK_TAP_TERM 100
+
+// Leader key
+// #define LEADER_KEY_STRICT_KEY_PROCESSING
+#define LEADER_TIMEOUT 190
 
 // Underglow configuration
 #ifdef RGBLIGHT_ENABLE
