+++ keyboards/ferris/keymaps/hbmorrisonv5/rules.mk
@@ -1,2 +1,4 @@
+TAP_DANCE_ENABLE = yes
 MOUSEKEY_ENABLE = yes
 SEND_STRING_ENABLE = yes
+LEADER_ENABLE = yes
